Created block for story hub
