+++ blocks/columns/columns.css
@@ -154,10 +154,6 @@
     right: 0px
   }
 
-  .columns-wrapper :is(.lg\:max-h-\[475px\]) {
-    max-height: 475px
-  }
-
   .columns-wrapper :is(.lg\:w-1\/2) {
     width: 50%
   }
