+++ blocks/footer/footer.css
@@ -259,11 +259,6 @@
     -moz-column-gap: 1.5rem;
          column-gap: 1.5rem
   }
-
-  .footer-wrapper :is(.md\:px-6) {
-    padding-left: 1.5rem;
-    padding-right: 1.5rem
-  }
 }
 
 @media (min-width: 1024px) {
@@ -284,11 +279,6 @@
 
   .footer-wrapper :is(.lg\:flex-row) {
     flex-direction: row
-  }
-
-  .footer-wrapper :is(.lg\:px-0) {
-    padding-left: 0px;
-    padding-right: 0px
   }
 
   .footer-wrapper :is(.lg\:px-6) {
