+++ blocks/story-hub/_story-hub.json
@@ -1,0 +1,22 @@
+{
+    "definitions": [
+      {
+        "title": "Story Hub",
+        "id": "story-hub",
+        "plugins": {
+          "xwalk": {
+            "page": {
+              "resourceType": "core/franklin/components/block/v1/block",
+              "template": {
+                "name": "Story Hub",
+                "model": "story-hub"
+              }
+            }
+          }
+        }
+      }
+    ],
+    "models": [],
+    "filters": []
+  }
+
