+++ component-definition.json
@@ -581,6 +581,21 @@
           }
         },
         {
+          "title": "Story Hub",
+          "id": "story-hub",
+          "plugins": {
+            "xwalk": {
+              "page": {
+                "resourceType": "core/franklin/components/block/v1/block",
+                "template": {
+                  "name": "Story Hub",
+                  "model": "story-hub"
+                }
+              }
+            }
+          }
+        },
+        {
           "title": "Teasers",
           "id": "teasers",
           "plugins": {
