+++ component-filters.json
@@ -39,7 +39,8 @@
       "related-stories",
       "dynamic-cards",
       "form",
-      "carousel"
+      "carousel",
+      "story-hub"
     ]
   },
   {
