+++ models/_section.json
@@ -92,7 +92,8 @@
         "related-stories",
         "dynamic-cards",
         "form",
-        "carousel"
+        "carousel",
+        "story-hub"
       ]
     }
   ]
